refactor: move SpinalSimRamAccessTester
